* develop:
  Changed Z to uint16_t outside of OTI, fixes #7

+++ include/parameters/fec_parameters.h
@@ -42,22 +42,23 @@ public:
 
     static fec_parameters derive_parameters(uint64_t data_len, uint16_t pay_len,
                                             uint64_t max_db_mem);
-    static fec_parameters new_local_instance(uint64_t F, uint16_t T, uint8_t Z,
+    static fec_parameters new_local_instance(uint64_t F, uint16_t T, uint16_t Z,
             uint16_t N, uint8_t Al);
 
     uint64_t data_length();
     uint16_t symbol_size();
-    uint8_t num_source_blocks();
+    uint16_t num_source_blocks();
     uint16_t interleaver_length();
     uint8_t symbol_alignment();
     uint16_t total_symbols();
 
 private:
-    static uint8_t derive_Z(uint16_t Kt, uint64_t WS, uint16_t T, uint8_t Al,
-                            uint16_t topN);
+    static uint16_t derive_Z(uint32_t Kt, uint64_t WS, uint16_t T, uint8_t Al,
+                             uint16_t topN);
 
-    static int derive_N(uint16_t Kt, uint8_t Z, uint64_t WS, uint16_t T, uint8_t Al,
-                        uint16_t topN);
+    static uint16_t derive_N(uint32_t Kt, uint16_t Z, uint64_t WS, uint16_t T,
+                             uint8_t Al,
+                             uint16_t topN);
 
     fec_parameters(uint64_t common_fec_oti, uint32_t scheme_spec_fec_oti);
     uint64_t common_fec_oti_;
